requirements.txt
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
bleach==3.3.0
boto3==1.17.70
botocore==1.20.70
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.2
django-cors-headers==3.7.0
django-heroku==0.0.0
django-storages==1.11.1
djangochannelsrestframework==0.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lockfile==0.12.2
luigi==3.0.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.10.1
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.8.0
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-daemon==2.3.0
python-dateutil==2.8.1
python-environ==0.4.54
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
s3transfer==0.4.2
scipy==1.6.1
seaborn==0.11.1
Send2Trash==1.5.0
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.1
terminado==0.9.4
testpath==0.4.4
timestring==1.6.4
tornado==6.1
traitlets==5.0.5
twilio==6.58.0
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.4
virtualenv==20.4.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
xlrd==2.0.1
zope.interface==5.4.0
